requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
cssselect2==0.7.0
Django==5.1.1
et-xmlfile==1.1.0
fonttools==4.54.1
html5lib==1.1
idna==3.10
lxml==5.3.0
openpyxl==3.1.5
oscrypto==1.3.0
pdfkit==1.0.0
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pypandoc==1.13
pypdf==5.0.1
PyPDF2==3.0.1
pyphen==0.16.0
python-bidi==0.6.0
python-docx==1.1.2
python-pptx==1.0.2
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.4
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
weasyprint==62.3
webencodings==0.5.1
xhtml2pdf==0.2.16
XlsxWriter==3.2.0
zopfli==0.2.3
